Modified   Source/DaisySeedHAL.cpp Modified   DSP Modified   DaisySeedFramework.vcxproj Modified   DaisySeedFramework.vcxproj.filters Modified   Include/DaisySeedFramework/ChunkDataAssembler.h Modified   Include/DaisySeedFramework/DaisySeedHAL.h Modified   Include/DaisySeedFramework/StaticString.h Modified   Include/DaisySeedFramework/StringUtils.h Modified   Include/DaisySeedFramework/WindowsHAL.h Modified   Source/SDL_HAL.cpp Modified   Source/StringUtils.cpp Modified   Source/USB/DaisyUSBAMCInterface.cpp Modified   Source/USB/DaisyUSBCDCInterface.cpp Modified   Source/USB/WindowsUSBCDCInterface.cpp Modified   Source/WindowsHAL.cpp Added      Include/DaisySeedFramework/WindowsInclude.h

+++ Source/USB/DaisyUSBAMCInterface.cpp
@@ -626,7 +626,7 @@ void DaisyUSBAMCInterface::BuildStreamingInterface(EP0Buffer& EP0Buffer, uint16_
 	ep->bLength = sizeof(USBEndpointDescriptor);
 	ep->bDescriptorType = USBDescTypes::Endpoint;
 	ep->bEndpointAddress = Endpoint;
-	ep->bmAttributes = (uint8_t)USBEndpointAttributes::Isochronous | (uint8_t)EndpointSyncTypes::Async;
+	ep->bmAttributes = (uint8_t)(USBEndpointAttributes::Isochronous | EndpointSyncTypes::Async);
 	ep->wMaxPacketSize = CalculateMaxPacketSize(ChannelCount, Config);
 	ep->bInterval = 1;
 	BufferOffset += sizeof(USBEndpointDescriptor);
